backend/requirements.txt
flask
flask-cors
dotenv
supabase
gunicorn
psycopg2-binary